Introduce generate.py script
Introduce `generate.py` script to generate build and run scripts. It
does so based on default configuration files (for building), KraftKit
configuration and runtime configuration files (for running).

Default configuration files are to be stored in the `defconfigs/`
directory of each app / example. KraftKit configuration is the the
`kraft.yaml` file of each app / example. Runtime configuration is the
`scripts/run.yaml` file.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>